chore(conformance): bless TS goldens for the protocol RPC shape
Bumps the comline-codegen-typescript pin to comline-typescript c535f14
(IR_HASH + params interfaces + discriminated-union error types + async
methods) and re-blesses:

  - protocol/typescript.ts — the full new shape
  - primitives/typescript.ts — `u128` now maps to `bigint`

The `protocol` fixture's `IR_HASH` matches the Rust golden's exactly —
the canonical `schema_ir_hash` is language-neutral.

+++ conformance/tests/golden/primitives/typescript.ts
@@ -5,7 +5,7 @@ export interface Scalars {
     small: number;
     medium: number;
     large: number;
-    huge: u128;
+    huge: bigint;
     signed_small: number;
     signed_large: number;
     name: string;
